requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-users==14.0.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
makefun==1.15.6
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2==2.9.10
pwdlib==0.2.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
